% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dict}
\alias{data_dict}
\title{Data dictionary for uhreg data.}
\format{
A data frame with 3 variables:
\describe{
\item{item}{\code{chr} Name of the variable.}
\item{description}{\code{chr} Description of the variable.}
\item{value}{\code{list} Mapping of numeric values to character strings for
categorical variables. \code{NULL} for numeric variables.}
}
}
\usage{
data_dict
}
\description{
Data dictionary for uhreg data.
}
\keyword{datasets}
