% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_score.R
\name{discretize_score}
\alias{discretize_score}
\title{Discretize scores}
\usage{
discretize_score(x, name)
}
\arguments{
\item{x}{A numeric vector.}

\item{name}{The name of the score. One of \code{c("MINITQ_score", "TBF12_score", "THI_score", "TQ_score")}.}
}
\value{
A factor.
}
\description{
\code{discretize_score} disretizes a numeric questionnaire score according to
predefined cutoff values from literature.
}
\examples{
x <- uhreg$MINITQ_score
discretize_score(x, "MINITQ_score")

}
