% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uhreg}
\alias{uhreg}
\title{University of Regensburg (UHREG) tinnitus data.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2299 rows and 295 columns.
}
\usage{
uhreg
}
\description{
University of Regensburg (UHREG) tinnitus data.
}
\keyword{datasets}
